pyproject.toml
[tool.poetry]
name = "jg.coop"
version = "0.0"
description = "The junior.guru website, Discord bot, and synchronization scripts"
authors = ["Honza Javorek <honza@junior.guru>"]
license = "MIT"
readme = "README.md"
homepage = "https://junior.guru/"
repository = "https://github.com/juniorguru/junior.guru/"
packages = [{ include = "jg" }]

[tool.poetry.scripts]
jg = "jg.coop.cli:main"
coop = "jg.coop.cli:main"

[tool.poetry.dependencies]
python = "3.11.*"
jinja2 = "3.1.3"
gspread = "6.1.0"
oauth2client = "4.1.3"
arrow = "1.3.0"
markdown = "3.6"
flask = "3.0.3"
frozen-flask = "1.0.2"
pillow = "10.3.0"
peewee = "3.17.1"
strictyaml = "1.7.3"
lxml = "4.9.3"
cssselect = "1.2.0"
google-api-python-client = "2.125.0"
requests = "2.31.0"
fiobank = "3.0.0"
emoji = "2.11.0"
favicon = "0.7.0"
feedparser = "6.0.11"
py-cord = {version = "2.5.0", extras = ["speed"]}
gql = {extras = ["requests"], version = "3.5.0"}
mkdocs = "1.5.3"
markupsafe = "2.1.5"
ics = {version = "0.8.0.dev0", allow-prereleases = true}
pod2gen = "1.0.3"
python-slugify = "8.0.4"
playwright = "1.43.0"
click = "8.1.7"
mkdocs-gen-files = "0.5.0"
mkdocs-awesome-pages-plugin = "2.9.2"
"jg.chick" = { git = "https://github.com/juniorguru/chick.git" }
pycircleci = "0.7.0"
livereload = "2.6.3"
beautifulsoup4 = "4.12.3"
czech-sort = "1.1.0"
teemup = "1.0.1"
stripe = "8.10.0"
diskcache = "5.6.3"
apify-client = "1.6.4"
apify-shared = "1.1.1"
openai = "1.16.1"
tenacity = "8.2.3"
lingua-language-detector = "2.0.2"
pygithub = "2.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "8.1.1"
pync = "2.0.3"
sqlite-utils = "3.36"
ghp-import = "2.1.0"
pytest-asyncio = "0.23.6"
scour = "0.38.2"
pytest-ruff = "0.3.1"
ruff = "0.3.1"

[tool.pytest.ini_options]
python_files = "test_*.py"
testpaths = "tests"
norecursedirs = "node_modules"
addopts = "--ff --ruff --ruff-format"
filterwarnings = [
    "ignore:'audioop' is deprecated:DeprecationWarning",  # py-cord
]

[tool.ruff]
target-version = "py311"
extend-exclude = ["*_legacy"]

[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.lint.isort]
combine-as-imports = true
lines-after-imports = 2
known-third-party = ["jg.chick", "jg.hen"]
known-local-folder = ["testing_utils"]

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"
